+++ consensus/transaction_validator/transaction_validator.h
@@ -21,12 +21,12 @@ extern "C" {
 
 typedef struct transaction_validator_t {
   uint64_t snapshot_timestamp_ms;
-  uint16_t mwm;
+  uint8_t mwm;
 } transaction_validator_t;
 
 extern retcode_t iota_consensus_transaction_validator_init(
     transaction_validator_t *const tv, uint64_t snapshot_timestamp_ms,
-    uint16_t mwm);
+    uint8_t mwm);
 
 extern retcode_t iota_consensus_transaction_validator_destroy(
     transaction_validator_t *const tv);
